Extract Rekor feeder impl
